# openapi.yaml
openapi: 3.0.3
info:
  title: dotti.work API
  version: 2.0.0
  description: >
    API PHP nativa para descoberta de projetos open source com OAuth GitHub,
    perfil tecnico, preferencias, cache de repositorios/issues e matching
    deterministico. Todas as rotas listadas tambem sao registradas com o
    prefixo /api/v1.
servers:
  - url: https://api.dottiwork.com
    description: Producao
  - url: https://api.dottiwork.com/api/v1
    description: Producao com prefixo versionado
  - url: http://localhost/dottiwork_api/api
    description: Local
  - url: http://localhost/dottiwork_api/api/api/v1
    description: Local com prefixo versionado
...
paths:
...
  /repositories/{owner}/{repo}:
    get:
      tags: [Repositorios]
      operationId: getRepository
      summary: Detalhe de repositorio
      security:
        - BearerAuth: []
        - CookieAuth: []
      parameters:
        - $ref: "#/components/parameters/Owner"
        - $ref: "#/components/parameters/Repo"
      responses:
        "200":
          description: Repositorio, saude, estado do usuario e match quando existir.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RepositoryDetailResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "502":
          $ref: "#/components/responses/BadGateway"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque-session-token
      description: Token local opaco da API. O token GitHub nunca e exposto.
    CookieAuth:
      type: apiKey
      in: cookie
      name: dotti_session
      description: Cookie HttpOnly/Secure/SameSite=Lax usado pelo front-end com credentials include.
  parameters:
    Owner:
      in: path
      name: owner
      required: true
      schema:
        type: string
      example: facebook
    Repo:
      in: path
      name: repo
      required: true
      schema:
        type: string
      example: react
...
  responses:
...
    Unauthorized:
      description: Sessao ausente, invalida ou expirada.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    BadGateway:
      description: Falha ao consultar servico externo.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
...
    ErrorResponse:
      type: object
      required: [success, error]
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          required: [code, message, details]
          properties:
            code:
              type: string
              example: VALIDATION_ERROR
            message:
              type: string
              example: Dados invalidos.
            details:
              type: array
              items:
                $ref: "#/components/schemas/ErrorDetail"
    ErrorDetail:
      type: object
      additionalProperties: true
      properties:
        field:
          type: string
          example: technologies.0.technology_id
        message:
          type: string
          example: technology_id deve ser inteiro.
...
    Match:
      type: object
      properties:
        github_repository_id:
          type: integer
          format: int64
          nullable: true
        score:
          type: number
          format: float
          example: 92.5
        recommended_seniority:
          type: string
          enum: [junior, mid, senior]
        breakdown:
          $ref: "#/components/schemas/MatchBreakdown"
        reasons:
          type: array
          items:
            type: string
        match_reasons:
          type: array
          items:
            type: string
        shared_technologies:
          type: array
          items:
            type: string
        positives:
          type: array
          items:
            type: string
        challenges:
          type: array
          items:
            type: string
        health_checklist:
          type: array
          items:
            $ref: "#/components/schemas/MatchHealthChecklistItem"
        generated_at:
          type: string
          nullable: true
        expires_at:
          type: string
          nullable: true
        cached:
          type: boolean
    MatchHealthChecklistItem:
      type: object
      required: [key, label, passed]
      properties:
        key:
          type: string
          example: has_readme
        label:
          type: string
          example: Has README
        passed:
          type: boolean
    MatchBreakdown:
      type: object
      properties:
        stack:
          type: number
        difficulty:
          type: number
        issues:
          type: number
        activity:
          type: number
        health:
          type: number
        contribution_readiness:
          type: number
    RepositoryDetailResponse:
      type: object
      required: [success, data]
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          required: [repository, health, user_state, match]
          properties:
            repository:
              $ref: "#/components/schemas/RepositorySummary"
            health:
              allOf:
                - $ref: "#/components/schemas/RepositoryHealth"
              nullable: true
            user_state:
              allOf:
                - $ref: "#/components/schemas/RepositoryStateValue"
              nullable: true
            match:
              allOf:
                - $ref: "#/components/schemas/Match"
              nullable: true
    RepositorySummary:
      type: object
      properties:
        github_repository_id:
          type: integer
          format: int64
          nullable: true
        owner:
          type: string
          nullable: true
        name:
          type: string
          nullable: true
        full_name:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        url:
          type: string
          format: uri
          nullable: true
        homepage_url:
          type: string
          format: uri
          nullable: true
        avatar_url:
          type: string
          format: uri
          nullable: true
        primary_language:
          type: string
          nullable: true
        languages:
          type: array
          items:
            type: string
        topics:
          type: array
          items:
            type: string
        stars:
          type: integer
        forks:
          type: integer
        watchers:
          type: integer
        open_issues:
          type: integer
        good_first_issues:
          type: integer
        help_wanted_issues:
          type: integer
        license:
          type: string
          nullable: true
        last_pushed_at:
          type: string
          nullable: true
        last_updated_at:
          type: string
          nullable: true
        project_size:
          type: string
          enum: [small, medium, large]
        activity_score:
          type: integer
          minimum: 0
          maximum: 100
        activity_label:
          type: string
          enum: [low, moderate, active, very_active]
        health_score:
          type: integer
          minimum: 0
          maximum: 100
          nullable: true
        html_url:
          type: string
          format: uri
          nullable: true
          deprecated: true
        updated_at:
          type: string
          nullable: true
          deprecated: true
        homepage:
          type: string
          format: uri
          nullable: true
          deprecated: true
    RepositoryHealth:
      type: object
      properties:
        score:
          type: integer
          minimum: 0
          maximum: 100
        has_readme:
          type: boolean
        has_contributing:
          type: boolean
        has_code_of_conduct:
          type: boolean
        has_ci:
          type: boolean
        has_tests:
          type: boolean
        has_contribution_labels:
          type: boolean
...
    RepositoryStateValue:
      type: string
      enum: [saved, ignored, researching, working, pull_request_sent, contributed, archived]
...